+++ app/build.gradle
@@ -43,6 +43,7 @@ repositories {
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
     testCompile 'junit:junit:4.12'
+    testCompile "org.robolectric:robolectric:3.1.2"
 
     compile "com.android.support:appcompat-v7:$supportLibVersion"
     compile "com.android.support:recyclerview-v7:$supportLibVersion"
@@ -61,10 +62,5 @@ dependencies {
     apt "frankiesardo:icepick-processor:$icePickVersion"
 
     compile "com.pushtorefresh.storio:sqlite:$storeioVersion"
-    compile "com.pushtorefresh.storio:sqlite-annotations:$storeioVersion"
     compile "com.pushtorefresh.storio:content-resolver:$storeioVersion"
-    compile "com.pushtorefresh.storio:content-resolver-annotations:$storeioVersion"
-
-    apt "com.pushtorefresh.storio:content-resolver-annotations-processor:$storeioVersion"
-    apt "com.pushtorefresh.storio:sqlite-annotations-processor:$storeioVersion"
 }
