--- app/src/androidTest/java/ru/yandex/yamblz/content_provider/ApplicationTest.java
@@ -1,15 +1,0 @@
-package ru.yandex.yamblz.content_provider;
-
-import android.app.Application;
-import android.test.ApplicationTestCase;
-
-/**
- * <a href="http://d.android.com/tools/testing/testing_android.html">Testing Fundamentals</a>
- */
-public class ApplicationTest extends ApplicationTestCase<Application>
-{
-    public ApplicationTest()
-    {
-        super(Application.class);
-    }
-}
