+++ app/src/androidTest/java/ru/yandex/yamblz/cp/ApplicationTest.java
@@ -1,0 +1,15 @@
+package ru.yandex.yamblz.cp;
+
+import android.app.Application;
+import android.test.ApplicationTestCase;
+
+/**
+ * <a href="http://d.android.com/tools/testing/testing_android.html">Testing Fundamentals</a>
+ */
+public class ApplicationTest extends ApplicationTestCase<Application>
+{
+    public ApplicationTest()
+    {
+        super(Application.class);
+    }
+}
