+++ app/src/main/AndroidManifest.xml
@@ -1,19 +1,19 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest
     xmlns:android="http://schemas.android.com/apk/res/android"
-    package="ru.yandex.yamblz.content_provider">
+    package="ru.yandex.yamblz.cp">
 
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
 
     <application
         android:allowBackup="false"
-        android:name=".App"
+        android:name="ru.yandex.yamblz.cp.App"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".dashboard.DashboardActivity">
+        <activity android:name="ru.yandex.yamblz.cp.dashboard.DashboardActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
 
@@ -23,7 +23,7 @@
 
         <provider
             android:authorities="ru.yandex.yamblz.artists_provider"
-            android:name=".data.repository.source.local.provider.ArtistsContentProvider"
+            android:name="ru.yandex.yamblz.cp.data.repository.source.local.provider.ArtistsContentProvider"
             android:enabled="true"
             android:exported="true"
             />
