--- app/src/main/java/ru/yandex/yamblz/content_provider/App.java
@@ -1,20 +1,0 @@
-package ru.yandex.yamblz.content_provider;
-
-import android.app.Application;
-
-import ru.yandex.yamblz.content_provider.data.repository.ArtistsRepository;
-import ru.yandex.yamblz.content_provider.data.repository.source.local.LocalDataSource;
-import ru.yandex.yamblz.content_provider.data.repository.source.remote.RemoteDataSource;
-
-/**
- * Created by platon on 02.08.2016.
- */
-public class App extends Application
-{
-    @Override
-    public void onCreate()
-    {
-        super.onCreate();
-        ArtistsRepository.init(new LocalDataSource(getContentResolver()), new RemoteDataSource());
-    }
-}
