--- app/src/main/java/ru/yandex/yamblz/content_provider/api/ArtistsApi.java
@@ -1,17 +1,0 @@
-package ru.yandex.yamblz.content_provider.api;
-
-
-import java.util.List;
-
-import retrofit2.http.GET;
-import ru.yandex.yamblz.content_provider.data.entity.Artist;
-import rx.Observable;
-
-/**
- * Created by postnov on 12.04.2016.
- */
-public interface ArtistsApi
-{
-    @GET("mobilization-2016/artists.json")
-    Observable<List<Artist>> getArtists();
-}
