--- app/src/main/java/ru/yandex/yamblz/content_provider/data/entity/Cover.java
@@ -1,36 +1,0 @@
-package ru.yandex.yamblz.content_provider.data.entity;
-
-import com.google.gson.annotations.SerializedName;
-
-import java.io.Serializable;
-
-public class Cover implements Serializable
-{
-    @SerializedName("small")
-    private String mCoverSmall;
-
-    @SerializedName("big")
-    private String mCoverBig;
-
-    public Cover(String small, String big)
-    {
-        mCoverSmall = small;
-        mCoverBig = big;
-    }
-
-    public String getSmall() {
-        return mCoverSmall;
-    }
-
-    public void setCoverSmall(String coverSmall) {
-        this.mCoverSmall = coverSmall;
-    }
-
-    public String getCoverBig() {
-        return mCoverBig;
-    }
-
-    public void setCoverBig(String coverBig) {
-        this.mCoverBig = coverBig;
-    }
-}
