--- app/src/main/java/ru/yandex/yamblz/content_provider/data/entity/mapper/TypeMapper.java
@@ -1,9 +1,0 @@
-package ru.yandex.yamblz.content_provider.data.entity.mapper;
-
-/**
- * Created by platon on 06.08.2016.
- */
-public interface TypeMapper<From, To>
-{
-    To map(From source);
-}
