--- app/src/main/java/ru/yandex/yamblz/content_provider/data/repository/DataSource.java
@@ -1,16 +1,0 @@
-package ru.yandex.yamblz.content_provider.data.repository;
-
-import java.util.List;
-import ru.yandex.yamblz.content_provider.data.entity.Artist;
-import rx.Observable;
-
-/**
- * Created by platon on 31.07.2016.
- */
-public interface DataSource
-{
-    Observable<List<Artist>> artists();
-    void put(List<Artist> artists);
-    void delete();
-    int size();
-}
