+++ app/src/main/java/ru/yandex/yamblz/cp/data/entity/mapper/TypeMapper.java
@@ -1,0 +1,9 @@
+package ru.yandex.yamblz.cp.data.entity.mapper;
+
+/**
+ * Created by platon on 06.08.2016.
+ */
+public interface TypeMapper<From, To>
+{
+    To map(From source);
+}
