+++ app/src/main/java/ru/yandex/yamblz/cp/data/exception/ErrorBundle.java
@@ -1,0 +1,11 @@
+package ru.yandex.yamblz.cp.data.exception;
+
+/**
+ * Created by platon on 08.08.2016.
+ */
+public interface ErrorBundle
+{
+    Throwable getException();
+
+    String getMessage();
+}
