--- app/src/test/java/ru/yandex/yamblz/content_provider/ExampleUnitTest.java
@@ -1,17 +1,0 @@
-package ru.yandex.yamblz.content_provider;
-
-import org.junit.Test;
-
-import static org.junit.Assert.*;
-
-/**
- * To work on unit tests, switch the Test Artifact in the Build Variants view.
- */
-public class ExampleUnitTest
-{
-    @Test
-    public void addition_isCorrect() throws Exception
-    {
-        assertEquals(4, 2 + 2);
-    }
-}
